pyproject.toml
[project]
name = "quiz-generator"
version = "0.1.0"
description = "AI-powered quiz generator service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "openai",
  "pydantic",
  "pydantic-settings",
  "python-dotenv",
  "sqlmodel>=0.0.14",
  "psycopg[binary]>=3.1",
  "pgvector>=0.2.4",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["app*"]
